+++ python/sglang/srt/entrypoints/engine.py
@@ -37,7 +37,7 @@ setattr(threading, "_register_atexit", lambda *args, **kwargs: None)
 import torch
 import uvloop
 
-from sglang.srt.entrypoints.EngineBase import EngineBase
+from sglang.srt.entrypoints.engine_base import EngineBase
 from sglang.srt.managers.data_parallel_controller import (
     run_data_parallel_controller_process,
 )
