+++ python/sglang/srt/entrypoints/http_server_engine.py
@@ -1,17 +1,12 @@
-import copy
-import dataclasses
 import multiprocessing
-import pickle
-import threading
 import time
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import List, Optional, Tuple
 
 import pybase64
 import requests
 import torch
-import torch.distributed as dist
 
-from sglang.srt.entrypoints.EngineBase import EngineBase
+from sglang.srt.entrypoints.engine_base import EngineBase
 from sglang.srt.entrypoints.http_server import launch_server
 from sglang.srt.server_args import ServerArgs
 from sglang.srt.utils import MultiprocessingSerializer, kill_process_tree
